+++ src/conditions/base/contract.ts
@@ -6,6 +6,39 @@ import { RpcCondition, rpcConditionSchema } from './rpc';
 
 export const STANDARD_CONTRACT_TYPES = ['ERC20', 'ERC721'];
 
+const functionAbiVariable = Joi.object({
+  internalType: Joi.string().required(),
+  name: Joi.string().required(),
+  type: Joi.string().required(),
+});
+
+const functionAbiSchema = Joi.object({
+  name: Joi.string().required(),
+  type: Joi.string().valid('function').required(),
+  inputs: Joi.array().items(functionAbiVariable),
+  outputs: Joi.array().items(functionAbiVariable),
+  // TODO: Should we restrict this to 'view'?
+  // stateMutability: Joi.string().valid('view').required(),
+}).custom((functionAbi, helper) => {
+  // Validate method name
+  const method = helper.state.ancestors[0].method;
+  if (functionAbi.name !== method) {
+    return helper.message({
+      custom: '"method" must be the same as "functionAbi.name"',
+    });
+  }
+
+  // Validate nr of parameters
+  const parameters = helper.state.ancestors[0].parameters;
+  if (functionAbi.inputs?.length !== parameters.length) {
+    return helper.message({
+      custom: '"parameters" must have the same length as "functionAbi.inputs"',
+    });
+  }
+
+  return functionAbi;
+});
+
 const contractMethodSchemas: Record<string, Joi.Schema> = {
   ...rpcConditionSchema,
   contractAddress: Joi.string().pattern(ETH_ADDRESS_REGEXP).required(),
@@ -13,7 +46,7 @@ const contractMethodSchemas: Record<string, Joi.Schema> = {
     .valid(...STANDARD_CONTRACT_TYPES)
     .optional(),
   method: Joi.string().required(),
-  functionAbi: Joi.object().optional(),
+  functionAbi: functionAbiSchema.optional(),
   parameters: Joi.array().required(),
 };
